refactor: extract server-mates.js from server.js
Move mate message handlers into dedicated server-mates module:
mate_create, mate_list, mate_delete, mate_readd_builtin,
mate_list_available_builtins, mate_update. Also includes
refreshTeamSections and scheduleRegistryRefresh helpers.

server.js uses a forward reference pattern for mateHandler since
addProject/removeProject call scheduleRegistryRefresh but are
defined before the mateHandler is initialized.

server.js: 1580 -> 1343 lines. server-mates.js: 273 lines.
